pyproject.toml
[tool.poetry]
name = "task-list-api"
version = "0.1.0"
description = ""
authors = ["Yen-Chen Kuo <yenchenkuo9331@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.4"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic-settings = "^2.3.3"
pymysql = "^1.1.1"
cryptography = "^42.0.8"
pyjwt = "^2.8.0"
pytest-cov = "^5.0.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
